+++ sources/wavelet_buffer.cc
@@ -5,18 +5,16 @@
 
 #include <functional>
 #include <memory>
-#include <sstream>
 #include <string>
 #include <tuple>
 #include <utility>
 #include <vector>
-
-#include <cereal/archives/binary.hpp>
+#include <iostream>
 
 #include "wavelet_buffer/wavelet_buffer.h"
 #include "wavelet_buffer/wavelet_buffer_view.h"
 
-namespace drift::dsp {
+namespace drift {
 
 using drift::sf::SfCompressor;
 
@@ -527,4 +525,4 @@ std::ostream& operator<<(std::ostream& os, const WaveletBuffer& wb) {
   return os;
 }
 
-}  // namespace drift::dsp
+}  // namespace drift