+++ sources/wavelet_buffer_view.cc
@@ -2,10 +2,11 @@
 #include "wavelet_buffer/wavelet_buffer_view.h"
 
 #include <memory>
+#include <iostream>
 
 #include "wavelet_buffer/wavelet_utils.h"
 
-namespace drift::dsp {
+namespace drift {
 
 class WaveletBufferView::Impl {
  public:
@@ -96,4 +97,4 @@ WaveletBufferView::operator WaveletBuffer() const {
   return impl_->CastToBuffer();
 }
 
-}  // namespace drift::dsp
+}  // namespace drift